Do not send host shortcut sequences to guest OS
When raw keyboard input is enabled, send the keyboard sequences to the
guest OS only if the shortcut sequence is not used by the host. Sending
the shortcut sequence processed by host sends irregular keyboard
patterns to the guest OS, such as sending key press events but not key
release events for the shortcut sequences.

Bug: b/369904951
Change-Id: Ia2e64504264aa9fc2627924acc601d1438f1999c